Index inlined functions by abstract origin
An individual function may be inlined many different times within a
single program. While version 5 of the DWARF specification describes a
`.debug_names` section which allows efficiently mapping from the linkage
name of a function to all of the sites where it is inlined (among other
things), we ultimately decided we could not rely on the presence of this
section [1]. As such, when we parse the DWARF data for a binary, we
build our own index of inlined functions, mapping from the DWARF DIE
address of each "abstract instance root" to the DWARF DIE address of
each "concrete inlined instance entry" (i.e. each site where the
function was inlined) [2].

Because modern optimizing compilers are _extremely_ aggressive about
inlining, there may be a very large number of inlined functions. In
order to accommodate this without dramatically increasing the amount of
time indexing takes, the inlined function index is formed across
multiple thread-local maps which are unioned together at the end of
indexing (a single global map with a mutex proved to be a major
bottleneck, hence this approach differing from that used for indexing
specifications).

[1] https://dwarfstd.org/doc/DWARF5.pdf#page=154
[2] https://dwarfstd.org/doc/DWARF5.pdf#page=101

Signed-off-by: Kevin Svetlitski <svetlitski@fb.com>