Add `drgn.Type.die_offset()` utility method
`drgn.Type.die_offset()` returns the offset of the DWARF DIE
corresponding to the given type. This is very useful during debugging
when one wants to inspect the relevant DWARF DIE, but dumping *all* of
the DWARF as text for large binaries takes a prohibitively long amount
of time. Knowing the offset of the relevant DIE allows one to instead
dump just the DWARF for that DIE and it's children much more cheaply by
running:
`llvm-dwarfdump -c --debug-info=OFFSET path/to/binary`.

Signed-off-by: Kevin Svetlitski <svetlitski@meta.com>